+++ src/components/Navigation/NavigationTabBar/WorkspacesTabButton.tsx
@@ -1,22 +1,15 @@
 import React from 'react';
-import type {OnyxCollection} from 'react-native-onyx';
 import type {ValueOf} from 'type-fest';
 import {PressableWithFeedback} from '@components/Pressable';
 import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
-import useOnyx from '@hooks/useOnyx';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWorkspaceNavigationRouteState from '@hooks/useWorkspaceNavigationRouteState';
 import useWorkspacesTabIndicatorStatus from '@hooks/useWorkspacesTabIndicatorStatus';
 import navigateToWorkspacesPage from '@libs/Navigation/helpers/navigateToWorkspacesPage';
-import type {DomainSplitNavigatorParamList, WorkspaceSplitNavigatorParamList} from '@navigation/types';
 import CONST from '@src/CONST';
-import NAVIGATORS from '@src/NAVIGATORS';
-import ONYXKEYS from '@src/ONYXKEYS';
-import type SCREENS from '@src/SCREENS';
-import type {Domain, Policy} from '@src/types/onyx';
 import NAVIGATION_TABS from './NAVIGATION_TABS';
 import TabBarItem from './TabBarItem';
 
@@ -33,29 +26,7 @@ function WorkspacesTabButton({selectedTab, isWideLayout}: WorkspacesTabButtonPro
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const {login: currentUserLogin} = useCurrentUserPersonalDetails();
 
-    const {lastWorkspacesTabNavigatorRoute, workspacesTabState, topmostFullScreenRoute} = useWorkspaceNavigationRouteState();
-    const params = workspacesTabState?.routes?.at(0)?.params as
-        | WorkspaceSplitNavigatorParamList[typeof SCREENS.WORKSPACE.INITIAL]
-        | DomainSplitNavigatorParamList[typeof SCREENS.DOMAIN.INITIAL];
-
-    const paramsPolicyID = params && 'policyID' in params ? params.policyID : undefined;
-    const paramsDomainAccountID = params && 'domainAccountID' in params ? params.domainAccountID : undefined;
-
-    const lastViewedPolicySelector = (policies: OnyxCollection<Policy>) => {
-        if (!lastWorkspacesTabNavigatorRoute || lastWorkspacesTabNavigatorRoute.name !== NAVIGATORS.WORKSPACE_SPLIT_NAVIGATOR || !paramsPolicyID) {
-            return undefined;
-        }
-        return policies?.[`${ONYXKEYS.COLLECTION.POLICY}${paramsPolicyID}`];
-    };
-    const [lastViewedPolicy] = useOnyx(ONYXKEYS.COLLECTION.POLICY, {selector: lastViewedPolicySelector}, [lastWorkspacesTabNavigatorRoute, paramsPolicyID]);
-
-    const lastViewedDomainSelector = (domains: OnyxCollection<Domain>) => {
-        if (!lastWorkspacesTabNavigatorRoute || lastWorkspacesTabNavigatorRoute.name !== NAVIGATORS.DOMAIN_SPLIT_NAVIGATOR || !paramsDomainAccountID) {
-            return undefined;
-        }
-        return domains?.[`${ONYXKEYS.COLLECTION.DOMAIN}${paramsDomainAccountID}`];
-    };
-    const [lastViewedDomain] = useOnyx(ONYXKEYS.COLLECTION.DOMAIN, {selector: lastViewedDomainSelector}, [lastWorkspacesTabNavigatorRoute, paramsDomainAccountID]);
+    const {lastWorkspacesTabNavigatorRoute, topmostFullScreenRoute, lastViewedPolicy, lastViewedDomain} = useWorkspaceNavigationRouteState();
 
     const navigateToWorkspaces = () => {
         navigateToWorkspacesPage({shouldUseNarrowLayout, currentUserLogin, policy: lastViewedPolicy, domain: lastViewedDomain, lastWorkspacesTabNavigatorRoute, topmostFullScreenRoute});
